Moved bme280.py to sensor_handler package; removed unnecessary class; setup main_driver.py; cleaned up variable names in hypsometric formula; added .gitignore;
